requirements.txt
fastapi
uvicorn
langchain
openai
requests
python-dotenv
serpapi
streamlit